Move code of playback panel buttons to the playback panel module.
