+++ kunquat/tracker/ui/views/playbackpanel.py
@@ -16,10 +16,6 @@ from PySide.QtGui import *
 
 import kunquat.tracker.cmdline as cmdline
 from .playbackposition import PlaybackPosition
-from .playbutton import PlayButton
-from .playpatternbutton import PlayPatternButton
-from .recordbutton import RecordButton
-from .silencebutton import SilenceButton
 from .updater import Updater
 
 
@@ -56,6 +52,38 @@ class PlaybackPanel(QToolBar, Updater):
         self.addWidget(self._interactivity_button)
 
 
+class PlayButton(QToolButton, Updater):
+
+    def __init__(self):
+        super().__init__()
+        self.setText('Play')
+        self.setToolTip('Play (Comma)')
+        self.setAutoRaise(True)
+
+    def _on_setup(self):
+        icon_bank = self._ui_model.get_icon_bank()
+        icon_path = icon_bank.get_icon_path('play')
+        icon = QIcon(icon_path)
+        self.setIcon(icon)
+        QObject.connect(self, SIGNAL('clicked()'), self._ui_model.play)
+
+
+class PlayPatternButton(QToolButton, Updater):
+
+    def __init__(self):
+        super().__init__()
+        self.setText('Play Pattern')
+        self.setToolTip('Play Pattern (Ctrl + Comma)')
+        self.setAutoRaise(True)
+
+    def _on_setup(self):
+        icon_bank = self._ui_model.get_icon_bank()
+        icon_path = icon_bank.get_icon_path('play_pattern')
+        icon = QIcon(icon_path)
+        self.setIcon(icon)
+        QObject.connect(self, SIGNAL('clicked()'), self._ui_model.play_pattern)
+
+
 class PlayFromCursorButton(QToolButton, Updater):
 
     def __init__(self):
@@ -70,6 +98,71 @@ class PlayFromCursorButton(QToolButton, Updater):
         icon = QIcon(icon_path)
         self.setIcon(icon)
         QObject.connect(self, SIGNAL('clicked()'), self._ui_model.play_from_cursor)
+
+
+class RecordButton(QToolButton, Updater):
+
+    def __init__(self):
+        super().__init__()
+        self._sheet_manager = None
+        self._playback_manager = None
+
+        self.setCheckable(True)
+        self.setText('Record')
+        self.setAutoRaise(True)
+
+    def _on_setup(self):
+        self.register_action('signal_record_mode', self._update_checked)
+
+        self._sheet_manager = self._ui_model.get_sheet_manager()
+        self._playback_manager = self._ui_model.get_playback_manager()
+
+        icon_bank = self._ui_model.get_icon_bank()
+        icon_path = icon_bank.get_icon_path('record')
+        icon = QIcon(icon_path)
+        self.setIcon(icon)
+
+        QObject.connect(self, SIGNAL('clicked()'),
+                        self._clicked)
+
+    def _update_checked(self):
+        old_block = self.blockSignals(True)
+        is_checked = self._playback_manager.is_recording()
+        self.setChecked(is_checked)
+        self.blockSignals(old_block)
+
+    def _clicked(self):
+        if self._playback_manager.is_recording():
+            self._playback_manager.stop_recording()
+        else:
+            self._playback_manager.start_recording()
+            self._sheet_manager.set_typewriter_connected(True)
+            self._ui_model.play()
+
+
+class SilenceButton(QToolButton, Updater):
+
+    def __init__(self):
+        super().__init__()
+        self._playback_manager = None
+
+        self.setText('Silence')
+        self.setToolTip('Silence (Period)')
+        self.setAutoRaise(True)
+
+    def _on_setup(self):
+        self._playback_manager = self._ui_model.get_playback_manager()
+
+        icon_bank = self._ui_model.get_icon_bank()
+        icon_path = icon_bank.get_icon_path('silence')
+        icon = QIcon(icon_path)
+        self.setIcon(icon)
+
+        QObject.connect(self, SIGNAL('clicked()'), self._clicked)
+
+    def _clicked(self):
+        self._playback_manager.stop_recording()
+        self._ui_model.silence()
 
 
 class InteractivityButton(QToolButton, Updater):
